削除機能、ページネーション機能搭載

+++ like-Insta/src/database/migrations/2019_03_09_130203_create__likes_table.php
@@ -15,8 +15,8 @@ class CreateLikesTable extends Migration
     {
         Schema::create('likes', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('post_id');
-            $table->integer('like_user_id');
+            $table->bigInteger('post_id');
+            $table->bigInteger('like_user_id');
             $table->timestamps();
         });
     }
